Move handles classes definition to single file
* Move OSInAppMessageAction to InAppMessage.js
* Move OSDeviceState, OSPermissionState, OSPermissionStateChanges, OSSubscriptionState, OSSubscriptionStateChanges, OSEmailSubscriptionState and OSEmailSubscriptionStateChanges to Subscription.js
* Move OSNotificationReceivedEvent, OSNotification, OSAndroidBackgroundImageLayout and OSActionButton to NotificationReceived.js
* Move OSNotificationOpenedResult and OSNotificationAction to NotificationOpened.js
